aws/examples/serverless-baby-agi/src/requirements.txt
jsonschema
awslambdaric
boto3
pandas
nltk
unstructured
wikipedia
duckdb
faiss-cpu
tiktoken
openai
duckduckgo-search==2.8.6
langchain==0.0.155